requirements.txt
absl-py==0.10.0
adbutils==0.9.0
aiowebsocket==1.0.0.dev2
alembic==1.4.2
aliyun-fc2==2.5.0
altgraph==0.17
amqp==2.6.0
aniso8601==8.0.0
apache-skywalking==0.1.0
apkutils2==1.0.0
appdirs==1.4.4
Appium-Python-Client==1.2.0
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.3.1
asn1crypto==1.3.0
astor==0.8.1
atomicwrites==1.3.0
attrs==20.3.0
Automat==20.2.0
backcall==0.1.0
backports.unittest-mock==1.5
baidubce==0.8.3
bce-python-sdk==0.8.61
beautifulsoup4==4.8.2
billiard==3.5.0.5
blinker==1.4
blurhash==1.1.4
boost==0.1
Brotli==1.0.7
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
cachelib==0.1
celery==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==4.0.0
charset-normalizer==2.0.3
cigam==0.0.3
click==7.1.2
colorama==0.4.3
comtypes==1.1.9
configparser==3.7.3
constantly==15.1.0
cryptography==2.4.2
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.8
DBUtils==1.3
decorator==4.4.2
demjson==2.2.4
Deprecated==1.2.11
deprecation==2.1.0
DingtalkChatbot==1.5.2
distlib==0.3.1
Django==3.1.4
dnspython==2.0.0
docopt==0.6.2
elasticsearch==7.13.0
enum-compat==0.0.3
esdk-obs-python==3.21.4
eventlet==0.25.2
fake-useragent==0.1.11
ffmpy==0.3.0
filelock==3.0.12
Flask==1.0.2
Flask-Admin==1.5.5
Flask-APScheduler==1.11.0
Flask-Compress==1.10.1
Flask-Cors==3.0.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
flask-zookeeper==0.0.8
fonttools==4.7.0
frida==15.0.18
frida-tools==10.2.1
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.34.0
grpcio-tools==1.34.0
gunicorn==19.8.1
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hdfs==2.5.8
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.7
importlib-metadata==1.7.0
importlib-resources==5.0.0
incremental==17.5.0
ipython==7.15.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.17.0
jieba==0.42.1
Jinja2==2.11.2
jiphy==1.2.2
jmespath==0.10.0
JPype1==1.3.0
jsonpath==0.82
kafka-python==2.0.1
kaitaistruct==0.8
kazoo==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.2.0
kombu==4.6.10
Lackey==0.7.3
ldap3==2.6.1
libs==0.0.9
logzero==1.6.3
lxml==4.5.1
Mako==1.1.3
Markdown==3.3.2
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.2.1
mitmproxy==5.0.0
mock==4.0.2
model-archiver==1.0.3
modelarts==1.1.3
monotonic==1.5
more-itertools==8.2.0
motor==2.1.0
mouse==0.7.1
MouseInfo==0.1.3
msgpack==1.0.2
mxnet-model-server==1.0.8
mysqlclient==1.4.6
natsort==7.1.1
networkx==2.5
numpy==1.18.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.8
pandas==1.0.4
parsel==1.6.0
parso==0.7.0
passlib==1.7.2
path==15.0.0
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==8.1.0
pluggy==0.13.1
prettytable==1.0.1
progress==1.5
prometheus-client==0.3.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.10.0
protocol==0.1.0
psutil==5.4.6
psycopg2==2.8.6
ptyprocess==0.6.0
publicsuffix2==2.20191221
py==1.8.1
py-eureka-client==0.7.4
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
pyDes==2.0.1
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==7.0.2
pyelftools==0.27
PyExecJS==1.5.1
PyGetWindow==0.0.9
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.6
pymongo==3.11.4
PyMsgBox==1.0.9
PyMySQL==0.9.3
pynput==1.7.2
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.8.1
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyquery==1.4.3
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
pytest==5.4.1
python-dateutil==2.8.1
python-editor==1.0.4
python-magic==0.4.15
PyTweening==1.0.3
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
pywinauto==0.6.8
pyxlsb==1.0.8
pyzbar==0.1.8
qiniu==7.2.8
queuelib==1.5.0
redis==3.2.0
reload==0.9
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
retry==0.9.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
Scrapy==2.4.1
scrapy-pipeline-mongodb==0.0.7
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.15.0
skywalking==0.1.0
sortedcontainers==2.1.0
soupsieve==2.0
SQLAlchemy==1.3.17
sqlparse==0.4.1
sseclient==0.0.27
style==1.1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testinfra==5.0.0
tldextract==2.2.3
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
Twisted==20.3.0
txmongo==19.2.0
typing-extensions==3.7.4.3
tzlocal==2.1
uiautomator2==2.12.3
update==0.0.1
urllib3==1.24.1
urwid==2.0.1
vine==1.3.0
virtualenv==20.3.0
w3lib==1.22.0
watchdog==0.10.2
wcwidth==0.2.3
websockets==8.1
Werkzeug==1.0.1
whichcraft==0.6.1
WMI==1.5.1
wrapt==1.12.1
wsproto==0.14.1
WTForms==2.3.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
XMind==1.2.0
xmltodict==0.12.0
you-get==0.4.1520
zipp==3.1.0
zope.interface==5.2.0
zstandard==0.12.0
